fix: resolve the final-round review's blocking and cleanup items
Consent scope (blocker 2): the certified-CMP gate keys on an explicit
EEA + UK + Switzerland country list (EU outermost regions included by
their own ISO codes) instead of continent Europe, which missed GDPR
territory on three other continents. enableUspr is scoped with the GPP
stub it depends on, and the spec now covers all four regimes: mandate,
outside it, US, unknown geo.

Preconditions in the map (blockers 3 and 4): slot 7's TODO now gates its
layoutKey on a visible ad label and a phone density re-measure with the
interval live; the Auto-ads anchor TODO requires the /articles URL-group
scoping and a real-phone check against the fixed footer bar before the
account switch flips.

Cleanup: the implementation moves to pages/articles/[id].tsx so the
X-Robots-Tag header fires on the shipping route and no SPA transition
can render the template under the legacy URL; comment rows keep one
stable element type so a moving interleave boundary cannot remount a
live comment subtree; expandSidebar leaves MainLayout with its last
consumer gone; the beforePopState reset carries its single-global-slot
caveat.
