+++ packages/shared/src/lib/geo.ts
@@ -45,3 +45,63 @@ export enum Continent {
 }
 
 export const outsideGdpr = ['US', 'IL'];
+
+/**
+ * Where Google requires a certified CMP (TCF): the EEA, the UK and
+ * Switzerland. Country codes rather than a continent, because the EU's
+ * outermost regions carry their own ISO codes on other continents (Réunion
+ * and Mayotte resolve to Africa, the Caribbean territories to the Americas)
+ * while continental Europe includes plenty of countries the mandate does not
+ * cover.
+ */
+const CERTIFIED_CMP_COUNTRIES = new Set([
+  // EU
+  'AT',
+  'BE',
+  'BG',
+  'HR',
+  'CY',
+  'CZ',
+  'DK',
+  'EE',
+  'FI',
+  'FR',
+  'DE',
+  'GR',
+  'HU',
+  'IE',
+  'IT',
+  'LV',
+  'LT',
+  'LU',
+  'MT',
+  'NL',
+  'PL',
+  'PT',
+  'RO',
+  'SK',
+  'SI',
+  'ES',
+  'SE',
+  // EEA outside the EU
+  'IS',
+  'LI',
+  'NO',
+  // UK and Switzerland
+  'GB',
+  'CH',
+  // EU member-state territories with their own ISO codes
+  'AX',
+  'GF',
+  'GP',
+  'MQ',
+  'RE',
+  'YT',
+  'BL',
+  'MF',
+  'PM',
+]);
+
+/** Unknown geo counts as covered: erring safe means asking, not serving. */
+export const requiresCertifiedCmp = (region?: string): boolean =>
+  !region || CERTIFIED_CMP_COUNTRIES.has(region.toUpperCase());
